refactor(cli): extract per-command files from main.ts
